+++ CHANGES.rst
@@ -5,6 +5,8 @@ threedi-qgis-plugin changelog
 0.9.3 (unreleased)
 ------------------
 
+- Add option to make graphs  ``absolute`` to the graph tool.
+
 - Bugfix predict calcualtion points: For endpoints always enumerate the
   ``last_seq_id`` by one.
 
@@ -14,10 +16,8 @@ threedi-qgis-plugin changelog
 0.9.2 (2017-02-14)
 ------------------
 
-
 - Adopted the column names for the ``predict_calc_points`` command to the
   newest 3Di migrations.
- 
 
 
 0.9.1 (2016-12-12)
@@ -32,7 +32,6 @@ threedi-qgis-plugin changelog
 
 0.9 (2016-11-28)
 ----------------
-
 
 - Update cumulative aggregation methods.